Rename module to cephfs-hadoop
