+++ underfs/cephfs-hadoop/src/main/resources/META-INF/services/alluxio.underfs.UnderFileSystemFactory
@@ -9,4 +9,4 @@
 # See the NOTICE file distributed with this work for information regarding copyright ownership.
 #
 
-alluxio.underfs.hcfscephfs.HcfsCephfsUnderFileSystemFactory
+alluxio.underfs.cephfshadoop.CephfsHadoopUnderFileSystemFactory
